# client/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Loom.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/access/session": {
      "post": {
        "tags": [
          "Access"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AccessTokenRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessSessionResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Access"
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/access/preview": {
      "post": {
        "tags": [
          "Access"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AccessTokenRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessLinkPreviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/access/claim": {
      "post": {
        "tags": [
          "Access"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AccessTokenRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "Role",
            "in": "query",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/UserRole"
                }
              ]
            }
          },
          {
            "name": "InstitutionId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Registered",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDir",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SafePage",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SafePageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SearchTerm",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponsePage"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{userId}": {
      "put": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AdminUpdateUserRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{userId}/role": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AdminSetRoleRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/coordinator-requests": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDir",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SafePage",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SafePageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SearchTerm",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoordinatorRequestResponsePage"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{userId}/reject-coordinator-request": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/coordinator-whitelist": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDir",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SafePage",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SafePageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SearchTerm",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoordinatorWhitelistEntryResponsePage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AddToWhitelistRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoordinatorWhitelistEntryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/coordinator-whitelist/{email}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessLinkPreviewResponse": {
        "required": [
          "academicYear",
          "canClaim",
          "exchangeGuid",
          "hasAccess",
          "partnerInstitutionName",
          "studentName"
        ],
        "type": "object",
        "properties": {
          "exchangeGuid": {
            "type": "string",
            "format": "uuid"
          },
          "studentName": {
            "type": "string"
          },
          "partnerInstitutionName": {
            "type": "string"
          },
          "academicYear": {
            "type": "string"
          },
          "hasAccess": {
            "type": "boolean"
          },
          "canClaim": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "AccessSessionResponse": {
        "required": [
          "exchangeGuid"
        ],
        "type": "object",
        "properties": {
          "exchangeGuid": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "AccessTokenRequest": {
        "required": [
          "token"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AddToWhitelistRequest": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AdminSetRoleRequest": {
        "required": [
          "role"
        ],
        "type": "object",
        "properties": {
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "AdminUpdateUserRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "jmbag": {
            "type": "string",
            "nullable": true
          },
          "mentor": {
            "type": "string",
            "nullable": true
          },
          "coordinatorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "institutionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AuthMeResponse": {
        "required": [
          "coordinatorId",
          "coordinatorName",
          "coordinatorRequestStatus",
          "email",
          "id",
          "institutionId",
          "institutionName",
          "isOnboarded",
          "jmbag",
          "mentor",
          "name",
          "role"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "jmbag": {
            "type": "string",
            "nullable": true
          },
          "mentor": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "isOnboarded": {
            "type": "boolean"
          },
          "institutionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "institutionName": {
            "type": "string",
            "nullable": true
          },
          "coordinatorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "coordinatorName": {
            "type": "string",
            "nullable": true
          },
          "coordinatorRequestStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CoordinatorRequestStatus"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ClaimResponse": {
        "required": [
          "exchangeGuid"
        ],
        "type": "object",
        "properties": {
          "exchangeGuid": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "CoordinatorRequestResponse": {
        "required": [
          "email",
          "id",
          "institutionName",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "institutionName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CoordinatorRequestResponsePage": {
        "required": [
          "hasDeleted",
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CoordinatorRequestResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CoordinatorRequestStatus": {
        "enum": [
          "Pending",
          "Rejected"
        ],
        "type": "string"
      },
...
      "CoordinatorWhitelistEntryResponse": {
        "required": [
          "createdAt",
          "email",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "CoordinatorWhitelistEntryResponsePage": {
        "required": [
          "hasDeleted",
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CoordinatorWhitelistEntryResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "UserListResponse": {
        "required": [
          "coordinatorId",
          "coordinatorName",
          "coordinatorRequestStatus",
          "email",
          "id",
          "institutionCity",
          "institutionId",
          "institutionName",
          "isOnboarded",
          "jmbag",
          "mentor",
          "name",
          "role"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "institutionName": {
            "type": "string",
            "nullable": true
          },
          "institutionCity": {
            "type": "string",
            "nullable": true
          },
          "institutionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "coordinatorRequestStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CoordinatorRequestStatus"
              }
            ],
            "nullable": true
          },
          "isOnboarded": {
            "type": "boolean"
          },
          "jmbag": {
            "type": "string",
            "nullable": true
          },
          "mentor": {
            "type": "string",
            "nullable": true
          },
          "coordinatorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "coordinatorName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserListResponsePage": {
        "required": [
          "hasDeleted",
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserListResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "UserRole": {
        "enum": [
          "Student",
          "Coordinator",
          "Admin"
        ],
        "type": "string"
      },
...
    }
  },
...
}